Improve cell size calculation of ContentRowView
Large estimated widths work okay, but there's a weird artifact that flashes for a brief moment if you swipe to the right & let go to have the bounce-back animation. A more accurate estimate prevents this glitch.

Additionally, ContentSquareCell was de-prioritizing the wrong constraint, which led to the image fade-in transition potentially animating from a wrongly-sized frame (which was also caused by the large estimated cell width). Now it correctly de-prioritizes contentView-related constraints, so the actual content of the cell is the same & only the bottom-right contentView anchors break if necessary (i.e. during initial cell sizing).